feat: refactor income components architecture and enable CSV downloads for all users
- Refactor income components with better separation of concerns:
  - Create IncomeView as main container handling filters and state
  - Refactor IncomeList to only handle list view display
  - Rename LotSummaryView to IncomeTable for better naming consistency
- Remove admin restrictions from CSV export functionality
- Hide lot filter in summary view (table view) as it's not applicable
- Improve filtering behavior to show only relevant data when type is selected
- Better mobile experience with cleaner filter interface
